+++ src/commands/Admin/load.js
@@ -1,4 +1,6 @@
 const { Command, Stopwatch } = require('klasa');
+const { pathExists } = require('fs-nextra');
+const { join } = require('path');
 
 module.exports = class extends Command {
 
@@ -16,9 +18,8 @@ module.exports = class extends Command {
 
 	async run(message, [core, store, path]) {
 		path = (path.endsWith('.js') ? path : `${path}.js`).split(this.regExp);
-		core = Boolean(core);
 		const timer = new Stopwatch();
-		const piece = store.load(path, core);
+		const piece = await (core ? this.tryEach(store, path) : store.load(store.userDirectory, path));
 
 		try {
 			if (!piece) throw message.language.get('COMMAND_LOAD_FAIL');
@@ -26,7 +27,7 @@ module.exports = class extends Command {
 			if (this.client.shard) {
 				await this.client.shard.broadcastEval(`
 					if (this.shard.id !== ${this.client.shard.id}) {
-						const piece = this.${piece.store}.load(${JSON.stringify(path)}, ${core});
+						const piece = this.${piece.store}.load(${piece.directory}, ${JSON.stringify(path)});
 						if (piece) piece.init();
 					}
 				`);
@@ -36,6 +37,11 @@ module.exports = class extends Command {
 			timer.stop();
 			throw message.language.get('COMMAND_LOAD_ERROR', store.name, piece ? piece.name : path.join('/'), error);
 		}
+	}
+
+	async tryEach(store, path) {
+		for (const dir of store.coreDirectories) if (await pathExists(join(dir, ...path))) return store.load(dir, path);
+		return undefined;
 	}
 
 };