+++ src/lib/settings/Configuration.js
@@ -1,4 +1,4 @@
-const { isObject, deepClone, tryParse, toTitleCase, arraysStrictEquals, getDeepTypeName, objectToTuples } = require('../util/util');
+const { isObject, deepClone, toTitleCase, arraysStrictEquals, getDeepTypeName, objectToTuples } = require('../util/util');
 const SchemaFolder = require('./SchemaFolder');
 const SchemaPiece = require('./SchemaPiece');
 
@@ -81,8 +81,9 @@ class Configuration {
 		 */
 		Object.defineProperty(this, '_existsInDB', { value: null, writable: true });
 
-		Configuration._merge(data, this.gateway.schema);
-		for (const key of this.gateway.schema.keys()) this[key] = data[key];
+		const { defaults } = this.gateway;
+		for (const key of this.gateway.schema.keys()) this[key] = defaults[key];
+		this._patch(data);
 	}
 
 	/**
@@ -111,9 +112,7 @@ class Configuration {
 	 * @returns {Configuration}
 	 */
 	clone() {
-		const clone = this.gateway.Configuration._clone(this, this.gateway.schema);
-		clone.id = this.id;
-		return new this.gateway.Configuration(this.gateway, clone);
+		return new this.constructor(this.gateway, this);
 	}
 
 	/**
@@ -289,10 +288,9 @@ class Configuration {
 		if (folders.length) array.push('= Folders =', ...folders.sort(), '');
 		if (keysTypes.length) {
 			for (const keyType of keysTypes.sort()) {
-				keys[keyType].sort();
-				array.push(`= ${toTitleCase(keyType)}s =`);
-				for (const key of keys[keyType]) array.push(`${key.padEnd(longest)} :: ${this.resolveString(message, folder[key])}`);
-				array.push('');
+				array.push(`= ${toTitleCase(keyType)}s =`,
+					...keys[keyType].sort().map(key => `${key.padEnd(longest)} :: ${this.resolveString(message, folder[key])}`),
+					'');
 			}
 		}
 		return array.join('\n');
@@ -537,12 +535,8 @@ class Configuration {
 	 * @private
 	 */
 	_patch(data) {
-		const { schema } = this.gateway;
-		for (const [key, piece] of schema) {
-			const value = data[key];
-			if (value === undefined || value === null) continue;
-			this[key] = piece.type === 'Folder' ? Configuration._patch(this[key], value, piece) : value;
-		}
+		if (typeof data !== 'object' || data === null) return;
+		this.constructor._patch(this, data, this.gateway.schema);
 	}
 
 	/**
@@ -551,7 +545,7 @@ class Configuration {
 	 * @returns {ConfigurationJSON}
 	 */
 	toJSON() {
-		return Configuration._clone(this, this.gateway.schema);
+		return Object.assign({}, ...[...this.gateway.schema.keys()].map(key => ({ [key]: deepClone(this[key]) })));
 	}
 
 	/**
@@ -561,50 +555,6 @@ class Configuration {
 	 */
 	toString() {
 		return `Configuration(${this.gateway.type}:${this.id})`;
-	}
-
-	/**
-	 * Assign data to the Configuration.
-	 * @since 0.5.0
-	 * @param {Object} data The data contained in the group
-	 * @param {(SchemaFolder|SchemaPiece)} schema A SchemaFolder or a SchemaPiece instance
-	 * @returns {Object}
-	 * @private
-	 */
-	static _merge(data, schema) {
-		for (const [key, piece] of schema) {
-			if (piece.type === 'Folder') {
-				if (!data[key]) data[key] = {};
-				data[key] = Configuration._merge(data[key], piece);
-			} else if (typeof data[key] === 'undefined' || data[key] === null) {
-				data[key] = deepClone(piece.default);
-			} else if (piece.array) {
-				if (typeof data[key] === 'string') data[key] = tryParse(data[key]);
-				if (Array.isArray(data[key])) continue;
-				piece.client.emit('wtf',
-					new TypeError(`${piece.path} | Expected an array, null, or undefined. Got: ${Object.prototype.toString.call(data[key])}`));
-			}
-		}
-
-		return data;
-	}
-
-	/**
-	 * Clone configs.
-	 * @since 0.5.0
-	 * @param {Object} data The data to clone
-	 * @param {SchemaFolder} schema A SchemaFolder instance
-	 * @returns {Object}
-	 * @private
-	 */
-	static _clone(data, schema) {
-		const clone = {};
-
-		for (const [key, piece] of schema) {
-			clone[key] = piece.type === 'Folder' ? Configuration._clone(data[key], piece) : deepClone(data[key]);
-		}
-
-		return clone;
 	}
 
 	/**
@@ -620,7 +570,7 @@ class Configuration {
 		for (const [key, piece] of schema) {
 			const value = data[key];
 			if (value === undefined || value === null) continue;
-			inst[key] = piece.type === 'Folder' ? Configuration._patch(inst[key], value, piece) : value;
+			inst[key] = piece.type === 'Folder' ? this._patch(inst[key], value, piece) : deepClone(value);
 		}
 
 		return inst;