+++ src/lib/settings/SchemaFolder.js
@@ -70,8 +70,8 @@ class SchemaFolder extends Schema {
 	 */
 	get defaults() {
 		const defaults = {};
-		for (const [key, value] of this) {
-			defaults[key] = value.type === 'Folder' ? value.defaults : value.default;
+		for (const [key, value] of this.entries()) {
+			defaults[key] = value.type === 'Folder' ? value.defaults : deepClone(value.default);
 		}
 		return defaults;
 	}